requirements.txt
mysql-connector-python
flask
flask-socketio
prometheus-flask-exporter
